+++ src/platform.ts
@@ -53,6 +53,13 @@ import {
   IlluminanceMeasurementCluster,
   TemperatureMeasurementCluster,
   DeviceTypeDefinition,
+  RelativeHumidityMeasurementCluster,
+  Cluster,
+  BitSchema,
+  TypeFromPartialBitSchema,
+  Attributes,
+  Commands,
+  Events,
 } from 'matterbridge';
 import { EveHistory, EveHistoryCluster } from 'matterbridge/history';
 import { AnsiLogger, BLUE, CYAN, GREEN, LogLevel, TimestampFormat, YELLOW, db, debugStringify, dn, er, hk, idn, nf, or, rs, wr, zb } from 'matterbridge/logger';
@@ -64,7 +71,7 @@ import path from 'path';
 import { Shelly } from './shelly.js';
 import { DiscoveredDevice } from './mdnsScanner.js';
 import { ShellyDevice } from './shellyDevice.js';
-import { ShellyComponent, ShellyCoverComponent, ShellyLightComponent, ShellySwitchComponent } from './shellyComponent.js';
+import { isLightComponent, ShellyComponent, ShellyCoverComponent, ShellyLightComponent, ShellySwitchComponent } from './shellyComponent.js';
 import { ShellyData, ShellyDataType } from './shellyTypes.js';
 
 type ConfigDeviceIp = Record<string, string>;
@@ -105,6 +112,8 @@ export function isValidBoolean(value: any): value is boolean {
  * Checks if a value is a valid string.
  *
  * @param {any} value - The value to be checked.
+ * @param {number} minLength - The min string length (optional).
+ * @param {number} maxLength - The max string length (optional).
  * @returns {boolean} A boolean indicating whether the value is a valid string.
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -164,7 +173,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
   public storedDevices = new Map<ShellyDeviceId, DiscoveredDevice>();
   public shellyDevices = new Map<ShellyDeviceId, ShellyDevice>();
   public bridgedDevices = new Map<ShellyDeviceId, MatterbridgeDevice>();
-  public bluDevices = new Map<string, MatterbridgeDevice>();
+  public bluBridgedDevices = new Map<string, MatterbridgeDevice>();
 
   // NodeStorageManager
   private nodeStorageManager?: NodeStorageManager;
@@ -273,15 +282,85 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
               const mbDevice = new MatterbridgeDevice(definition, undefined, config.debug as boolean);
               mbDevice.createDefaultBridgedDeviceBasicInformationClusterServer(bthomeDevice.name, bthomeDevice.addr, 0xfff1, 'Shelly', bthomeDevice.model);
               mbDevice.createDefaultPowerSourceReplaceableBatteryClusterServer();
+              if (bthomeDevice.model === 'Shelly BLU DoorWindow') mbDevice.addDeviceType(DeviceTypes.LIGHT_SENSOR);
+              if (bthomeDevice.model === 'Shelly BLU HT') mbDevice.addDeviceType(DeviceTypes.HUMIDITY_SENSOR);
+              if (bthomeDevice.model === 'Shelly BLU HT') mbDevice.addDeviceType(DeviceTypes.LIGHT_SENSOR);
+              if (bthomeDevice.model === 'Shelly BLU HT') mbDevice.addDeviceType(DeviceTypes.GENERIC_SWITCH);
+              if (bthomeDevice.model === 'Shelly BLU Motion') mbDevice.addDeviceType(DeviceTypes.LIGHT_SENSOR);
               mbDevice.addRequiredClusterServers(mbDevice);
-              if (bthomeDevice.model === 'Shelly BLU DoorWindow') mbDevice.createDefaultIlluminanceMeasurementClusterServer();
-              if (bthomeDevice.model === 'Shelly BLU HT') mbDevice.createDefaultRelativeHumidityMeasurementClusterServer();
-              if (bthomeDevice.model === 'Shelly BLU HT') mbDevice.createDefaultSwitchClusterServer();
-              if (bthomeDevice.model === 'Shelly BLU Motion') mbDevice.createDefaultIlluminanceMeasurementClusterServer();
-              await this.registerDevice(mbDevice);
-              this.bluDevices.set(key, mbDevice);
+              try {
+                await this.registerDevice(mbDevice);
+                this.bluBridgedDevices.set(key, mbDevice);
+              } catch (error) {
+                this.log.error(
+                  `Shelly device ${hk}${device.id}${er} host ${zb}${device.host}${er} failed to register BLU device ${idn}${bthomeDevice.name}${er}: ${error instanceof Error ? error.message : error}`,
+                );
+              }
             }
           }
+          device.on('bthomedevice_update', (addr: string, rssi: number, last_updated_ts: number) => {
+            if (!isValidString(addr, 11) || !isValidNumber(rssi, -100, 0) || !isValidNumber(last_updated_ts)) return;
+            const blu = this.bluBridgedDevices.get(addr);
+            if (!blu) {
+              this.log.error(`Shelly device ${hk}${device.id}${er} host ${zb}${device.host}${er} sent an unknown BLU device address ${CYAN}${addr}${er}`);
+              return;
+            }
+            const lastUpdatedTime = new Date(last_updated_ts * 1000);
+            this.log.info(
+              `**BLU observer message for BLU device ${idn}${blu?.deviceName ?? addr}${rs}${nf}: rssi ${YELLOW}${rssi}${nf} last_updated ${YELLOW}${lastUpdatedTime.toLocaleTimeString()}${nf}`,
+            );
+          });
+          device.on('bthomesensor_update', (addr: string, sensor: string, value: ShellyDataType) => {
+            if (!isValidString(addr, 11) || !isValidString(sensor, 6)) return;
+            const blu = this.bluBridgedDevices.get(addr);
+            if (!blu) {
+              this.log.error(`Shelly device ${hk}${device.id}${er} host ${zb}${device.host}${er} sent an unknown BLU device address ${CYAN}${addr}${er}`);
+              return;
+            }
+            this.log.info(`**BLU observer message for BLU device ${idn}${blu?.deviceName ?? addr}${rs}${nf}: sensor ${YELLOW}${sensor}${nf} value ${YELLOW}${value}${nf}`);
+            if (blu && sensor === 'Battery' && isValidNumber(value, 0, 100)) {
+              this.setAttribute(blu, PowerSource.Complete, 'batPercentRemaining', value * 2, true);
+              if (value < 10) this.setAttribute(blu, PowerSource.Complete, 'batChargeLevel', PowerSource.BatChargeLevel.Critical, true);
+              else if (value < 20) this.setAttribute(blu, PowerSource.Complete, 'batChargeLevel', PowerSource.BatChargeLevel.Warning, true);
+              else this.setAttribute(blu, PowerSource.Complete, 'batChargeLevel', PowerSource.BatChargeLevel.Ok, true);
+            }
+            if (blu && sensor === 'Temperature' && isValidNumber(value, -100, 100)) {
+              this.setAttribute(blu, TemperatureMeasurementCluster, 'measuredValue', value * 100, true);
+            }
+            if (blu && sensor === 'Humidity' && isValidNumber(value, 0, 100)) {
+              this.setAttribute(blu, RelativeHumidityMeasurementCluster, 'measuredValue', value * 100, true);
+            }
+            if (blu && sensor === 'Illuminance' && isValidNumber(value, 0, 10000)) {
+              const matterLux = Math.round(Math.max(Math.min(10000 * Math.log10(value), 0xfffe), 0));
+              this.setAttribute(blu, IlluminanceMeasurementCluster, 'measuredValue', matterLux, true);
+            }
+            if (blu && sensor === 'Motion' && isValidBoolean(value)) {
+              this.setAttribute(blu, OccupancySensingCluster, 'occupancy', { occupied: value }, true);
+            }
+            if (blu && sensor === 'Contact' && isValidBoolean(value)) {
+              this.setAttribute(blu, BooleanStateCluster, 'stateValue', !value, true);
+            }
+          });
+          device.on('bthomesensor_event', (addr: string, sensor: string, event: string) => {
+            if (!isValidString(addr, 11) || !isValidString(sensor, 6)) return;
+            const blu = this.bluBridgedDevices.get(addr);
+            if (!blu) {
+              this.log.error(`Shelly device ${hk}${device.id}${er} host ${zb}${device.host}${er} sent an unknown BLU device address ${CYAN}${addr}${er}`);
+              return;
+            }
+            this.log.info(`**BLU observer message for BLU device ${idn}${blu?.deviceName ?? addr}${rs}${nf}: sensor ${YELLOW}${sensor}${nf} event ${YELLOW}${event}${nf}`);
+            if (blu && sensor === 'Button' && isValidString(event, 9)) {
+              if (event === 'single_push') {
+                this.triggerSwitchEvent(blu, 'Single', true);
+              }
+              if (event === 'double_push') {
+                this.triggerSwitchEvent(blu, 'Double', true);
+              }
+              if (event === 'long_push') {
+                this.triggerSwitchEvent(blu, 'Long', true);
+              }
+            }
+          });
         }
       }
 
@@ -321,7 +400,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
         let composed = false;
         if (component.name === 'Light' || component.name === 'Rgb') {
           const lightComponent = device.getComponent(key);
-          if (lightComponent) {
+          if (/* lightComponent &&*/ isLightComponent(lightComponent)) {
             // Set the device type
             let deviceType = DeviceTypes.ON_OFF_LIGHT;
             if (lightComponent.hasProperty('brightness')) deviceType = DeviceTypes.DIMMABLE_LIGHT;
@@ -407,7 +486,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
             });
           }
         } else if (component.name === 'Switch' || component.name === 'Relay') {
-          const switchComponent = device.getComponent(key) as ShellySwitchComponent;
+          const switchComponent = device.getComponent(key);
           if (switchComponent) {
             let deviceType = onOffSwitch;
             if (config.exposeSwitch === 'light') deviceType = DeviceTypes.ON_OFF_LIGHT;
@@ -635,7 +714,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
           }
         } else if (component.name === 'Vibration' && config.exposeVibration !== 'disabled') {
           const vibrationComponent = device.getComponent(key);
-          if (vibrationComponent?.hasProperty('vibration')) {
+          if (vibrationComponent?.hasProperty('vibration') && isValidBoolean(vibrationComponent.getValue('vibration'))) {
             const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.GENERIC_SWITCH], []);
             child.addClusterServer(mbDevice.getDefaultSwitchClusterServer());
             if (!composed) mbDevice.addFixedLabel('composed', component.name);
@@ -655,6 +734,30 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
             composed = true;
             // Add event handler
             tempComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
+              this.shellyUpdateHandler(mbDevice, device, component, property, value);
+            });
+          }
+          if (tempComponent?.hasProperty('tC') && isValidNumber(tempComponent.getValue('tC'), -100, 100)) {
+            const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.TEMPERATURE_SENSOR], []);
+            const matterTemp = Math.min(Math.max(Math.round((tempComponent.getValue('tC') as number) * 100), -10000), 10000);
+            child.addClusterServer(mbDevice.getDefaultTemperatureMeasurementClusterServer(matterTemp));
+            if (!composed) mbDevice.addFixedLabel('composed', component.name);
+            composed = true;
+            // Add event handler
+            tempComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
+              this.shellyUpdateHandler(mbDevice, device, component, property, value);
+            });
+          }
+        } else if (component.name === 'Humidity' && config.exposeHumidity !== 'disabled') {
+          const humidityComponent = device.getComponent(key);
+          if (humidityComponent?.hasProperty('rh') && isValidNumber(humidityComponent.getValue('rh'), 0, 100)) {
+            const child = mbDevice.addChildDeviceTypeWithClusterServer(key, [DeviceTypes.HUMIDITY_SENSOR], []);
+            const matterHumidity = Math.min(Math.max(Math.round((humidityComponent.getValue('rh') as number) * 100), -10000), 10000);
+            child.addClusterServer(mbDevice.getDefaultRelativeHumidityMeasurementClusterServer(matterHumidity));
+            if (!composed) mbDevice.addFixedLabel('composed', component.name);
+            composed = true;
+            // Add event handler
+            humidityComponent.on('update', (component: string, property: string, value: ShellyDataType) => {
               this.shellyUpdateHandler(mbDevice, device, component, property, value);
             });
           }
@@ -869,6 +972,41 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
   }
 
   // TODO remove and use matterbridge method
+  protected setAttribute<F extends BitSchema, SF extends TypeFromPartialBitSchema<F>, A extends Attributes, C extends Commands, E extends Events>(
+    endpoint: Endpoint,
+    cluster: Cluster<F, SF, A, C, E>,
+    attribute: string,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    value: any,
+    log = false,
+  ) {
+    const clusterServer = endpoint.getClusterServer(cluster);
+    if (!clusterServer) {
+      if (log) this.log.error(`setAttribute error: Cluster ${cluster.name} not found on endpoint ${endpoint.name}:${endpoint.number}`);
+      return;
+    }
+    if (!clusterServer.isAttributeSupportedByName(attribute)) {
+      if (log) this.log.error(`setAttribute error: Attribute ${attribute} not found on Cluster ${cluster.name} on endpoint ${endpoint.name}:${endpoint.number}`);
+      return;
+    }
+    const capitalizedAttributeName = attribute.charAt(0).toUpperCase() + attribute.slice(1);
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    if (!(clusterServer as any)[`set${capitalizedAttributeName}Attribute`]) {
+      if (log)
+        this.log.error(`setAttribute error: Setter set${capitalizedAttributeName}Attribute not found on Cluster ${cluster.name} on endpoint ${endpoint.name}:${endpoint.number}`);
+      return;
+    }
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-empty-object-type
+    const setter = (clusterServer as any)[`set${capitalizedAttributeName}Attribute`] as (value: any) => {};
+    setter(value);
+    if (log)
+      this.log.info(
+        `**${db}Update endpoint ${or}${endpoint.name}:${endpoint.number}${db} attribute ${hk}${cluster.name}.${capitalizedAttributeName}${db} ` +
+          `to ${YELLOW}${typeof value === 'object' ? debugStringify(value) : value}${db}`,
+      );
+  }
+
+  // TODO remove and use matterbridge method
   protected configureColorControlCluster(endpoint: Endpoint, hueSaturation: boolean, xy: boolean, colorTemperature: boolean, colorMode?: number) {
     endpoint.getClusterServer(ColorControlCluster)?.setFeatureMapAttribute({ hueSaturation, enhancedHue: false, colorLoop: false, xy, colorTemperature });
     endpoint.getClusterServer(ColorControlCluster)?.setColorCapabilitiesAttribute({ hueSaturation, enhancedHue: false, colorLoop: false, xy, colorTemperature });
@@ -887,7 +1025,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
   }
 
   // TODO remove and use matterbridge method
-  protected triggerSwitchEvent(endpoint: Endpoint, event: string) {
+  protected triggerSwitchEvent(endpoint: Endpoint, event: string, log = false) {
     if (['Single', 'Double', 'Long'].includes(event)) {
       const cluster = endpoint.getClusterServer(
         SwitchCluster.with(
@@ -898,7 +1036,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
         ),
       );
       if (!cluster || !cluster.getFeatureMapAttribute().momentarySwitch) {
-        this.log.error(`triggerSwitchEvent ${event} error: Switch cluster with MomentarySwitch not found on endpoint ${endpoint.name}:${endpoint.number}`);
+        if (log) this.log.error(`triggerSwitchEvent ${event} error: Switch cluster with MomentarySwitch not found on endpoint ${endpoint.name}:${endpoint.number}`);
         return;
       }
       if (event === 'Single') {
@@ -908,7 +1046,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
         cluster.triggerShortReleaseEvent({ previousPosition: 1 });
         cluster.setCurrentPositionAttribute(0);
         cluster.triggerMultiPressCompleteEvent({ previousPosition: 1, totalNumberOfPressesCounted: 1 });
-        this.log.debug(`Trigger 'Single press' event for ${endpoint.name}:${endpoint.number}`);
+        if (log) this.log.info(`**${db}Trigger endpoint ${or}${endpoint.name}:${endpoint.number}${db} event ${hk}${cluster.name}.SinglePress${db}`);
       }
       if (event === 'Double') {
         cluster.setCurrentPositionAttribute(1);
@@ -921,7 +1059,7 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
         cluster.setCurrentPositionAttribute(0);
         cluster.triggerShortReleaseEvent({ previousPosition: 1 });
         cluster.triggerMultiPressCompleteEvent({ previousPosition: 1, totalNumberOfPressesCounted: 2 });
-        this.log.debug(`Trigger 'Double press' event for ${endpoint.name}:${endpoint.number}`);
+        if (log) this.log.info(`**${db}Trigger endpoint ${or}${endpoint.name}:${endpoint.number}${db} event ${hk}${cluster.name}.DoublePress${db}`);
       }
       if (event === 'Long') {
         cluster.setCurrentPositionAttribute(1);
@@ -929,24 +1067,26 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
         cluster.triggerLongPressEvent({ newPosition: 1 });
         cluster.setCurrentPositionAttribute(0);
         cluster.triggerLongReleaseEvent({ previousPosition: 1 });
-        this.log.debug(`Trigger 'Long press' event for ${endpoint.name}:${endpoint.number}`);
+        if (log) this.log.info(`**${db}Trigger endpoint ${or}${endpoint.name}:${endpoint.number}${db} event ${hk}${cluster.name}.LongPress${db}`);
       }
     }
     if (['Press', 'Release'].includes(event)) {
       const cluster = endpoint.getClusterServer(Switch.Complete);
       if (!cluster || !cluster.getFeatureMapAttribute().latchingSwitch) {
-        this.log.error(`triggerSwitchEvent ${event} error: Switch cluster with LatchingSwitch not found on endpoint ${endpoint.name}:${endpoint.number}`);
+        if (log) this.log.error(`triggerSwitchEvent ${event} error: Switch cluster with LatchingSwitch not found on endpoint ${endpoint.name}:${endpoint.number}`);
         return;
       }
       if (event === 'Press') {
         cluster.setCurrentPositionAttribute(1);
+        if (log) this.log.info(`**${db}Update endpoint ${or}${endpoint.name}:${endpoint.number}${db} attribute ${hk}${cluster.name}.CurrentPosition${db} to ${YELLOW}1${db}`);
         if (cluster.triggerSwitchLatchedEvent) cluster.triggerSwitchLatchedEvent({ newPosition: 1 });
-        this.log.debug(`Trigger ${event} event for ${endpoint.name}:${endpoint.number}`);
+        if (log) this.log.info(`**${db}Trigger endpoint ${or}${endpoint.name}:${endpoint.number}${db} event ${hk}${cluster.name}.Press${db}`);
       }
       if (event === 'Release') {
         cluster.setCurrentPositionAttribute(0);
+        if (log) this.log.info(`**${db}Update endpoint ${or}${endpoint.name}:${endpoint.number}${db} attribute ${hk}${cluster.name}.CurrentPosition${db} to ${YELLOW}0${db}`);
         if (cluster.triggerSwitchLatchedEvent) cluster.triggerSwitchLatchedEvent({ newPosition: 0 });
-        this.log.debug(`Trigger ${event} event for ${endpoint.name}:${endpoint.number}`);
+        if (log) this.log.info(`**${db}Trigger endpoint ${or}${endpoint.name}:${endpoint.number}${db} event ${hk}${cluster.name}.Release${db}`);
       }
     }
   }
@@ -1306,7 +1446,8 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
         // Calculate power from current and voltage
         const voltage = shellyComponent.hasProperty('voltage') ? (shellyComponent.getValue('voltage') as number) : undefined;
         if (isValidNumber(voltage, 0)) {
-          const power = value * voltage;
+          let power = value * voltage;
+          power = Math.round(power * 10000) / 10000;
           if (isValidNumber(power, 0)) cluster?.setConsumptionAttribute(power);
           if (cluster) shellyDevice.log.info(`${db}Update endpoint ${or}${endpoint.number}${db} attribute ${hk}EveHistory-consumption${db} ${YELLOW}${power}${db}`);
         }