+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 If you like this project and find it useful, please consider giving it a star on GitHub at https://github.com/Luligu/matterbridge-shelly and sponsoring it.
 
+- [shelly]: Added support for shellyhtg3.
+
+## [0.8.2] - 2024-08-14
+
+### Added
+
+- [shelly]: Added support for BLU devices (beta).
+- [shelly]: Added PowerSource.BatChargeLevel.Warning state when the battery is below 20%.
+- [shelly]: Added PowerSource.BatChargeLevel.Critical state when the battery is below 10%.
+- [shelly]: Added component Humidity (humidity).
+- [shelly]: Added detection of sleeping devices Gen. 2 and 3.
+
 ## [0.8.1] - 2024-08-12
 
 ### Changed
@@ -19,9 +31,10 @@ If you like this project and find it useful, please consider giving it a star on
 - [plugin]: Added debugWs option to the config.
 - [plugin]: Enabled enableBleDiscover option in the config.
 - [shelly]: Added support for BLU devices (alpha stage).
-- [shelly]: Added strict type checking to updates from devices.
 
 ### Fixed
+
+- [shelly]: Added strict type checking to updates from devices to prevent validation error caused by unsupported devices or firmware.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="./yellow-button.png" alt="Buy me a coffee" width="120">